# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MedEat API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /medicines:
    get:
      operationId: listMedicines
      tags: [medicines]
      summary: List all medicines
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, paused, finished, all]
      responses:
        "200":
          description: List of medicines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medicine"
    post:
      operationId: createMedicine
      tags: [medicines]
      summary: Add a new medicine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMedicineBody"
      responses:
        "201":
          description: Created medicine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
  /medicines/{id}:
    get:
      operationId: getMedicine
      tags: [medicines]
      summary: Get a medicine by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Medicine details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
    patch:
      operationId: updateMedicine
      tags: [medicines]
      summary: Update a medicine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMedicineBody"
      responses:
        "200":
          description: Updated medicine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
    delete:
      operationId: deleteMedicine
      tags: [medicines]
      summary: Delete a medicine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /medicines/{id}/take:
    post:
      operationId: takeDose
      tags: [medicines]
      summary: Mark a dose as taken and decrement pill count
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TakeDoseBody"
      responses:
        "200":
          description: Updated medicine with dose log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TakeDoseResponse"
  /medicines/{id}/refill:
    post:
      operationId: refillMedicine
      tags: [medicines]
      summary: Log a refill and update pill count
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefillMedicineBody"
      responses:
        "200":
          description: Updated medicine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
  /dose-logs:
    get:
      operationId: listDoseLogs
      tags: [dose-logs]
      summary: List dose logs
      parameters:
        - name: medicineId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of dose logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DoseLog"
...
  /recipes/suggestions:
    get:
      operationId: getRecipeSuggestions
      tags: [recipes]
      summary: Get AI-generated recipe suggestions based on profile and remaining macros
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Recipe suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recipe"
...
components:
  schemas:
...
    Medicine:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        dose:
          type: string
        form:
          type: string
          enum: [tablet, capsule, syrup, injection, drops, patch, other]
        frequency:
          type: string
          enum: [daily, twice_daily, three_times_daily, every_n_hours, specific_days, as_needed]
        timesOfDay:
          type: array
          items:
            type: string
        pillCount:
          type: integer
        refillThreshold:
          type: integer
        status:
          type: string
          enum: [active, paused, finished]
        foodInstruction:
          type: string
          enum: [before_food, after_food, with_water, any]
        prescriptionExpiry:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, dose, form, frequency, timesOfDay, pillCount, refillThreshold, status, foodInstruction, createdAt]
    CreateMedicineBody:
      type: object
      properties:
        name:
          type: string
        dose:
          type: string
        form:
          type: string
          enum: [tablet, capsule, syrup, injection, drops, patch, other]
        frequency:
          type: string
          enum: [daily, twice_daily, three_times_daily, every_n_hours, specific_days, as_needed]
        timesOfDay:
          type: array
          items:
            type: string
        pillCount:
          type: integer
        refillThreshold:
          type: integer
        foodInstruction:
          type: string
          enum: [before_food, after_food, with_water, any]
        prescriptionExpiry:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
      required: [name, dose, form, frequency, timesOfDay, pillCount, refillThreshold, foodInstruction]
    UpdateMedicineBody:
      type: object
      properties:
        name:
          type: string
        dose:
          type: string
        form:
          type: string
          enum: [tablet, capsule, syrup, injection, drops, patch, other]
        frequency:
          type: string
          enum: [daily, twice_daily, three_times_daily, every_n_hours, specific_days, as_needed]
        timesOfDay:
          type: array
          items:
            type: string
        pillCount:
          type: integer
        refillThreshold:
          type: integer
        status:
          type: string
          enum: [active, paused, finished]
        foodInstruction:
          type: string
          enum: [before_food, after_food, with_water, any]
        prescriptionExpiry:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
    TakeDoseBody:
      type: object
      properties:
        status:
          type: string
          enum: [taken, skipped, missed]
        scheduledTime:
          type: string
      required: [status]
    TakeDoseResponse:
      type: object
      properties:
        medicine:
          $ref: "#/components/schemas/Medicine"
        doseLog:
          $ref: "#/components/schemas/DoseLog"
        refillAlert:
          type: boolean
      required: [medicine, doseLog, refillAlert]
    RefillMedicineBody:
      type: object
      properties:
        newCount:
          type: integer
      required: [newCount]
    DoseLog:
      type: object
      properties:
        id:
          type: integer
        medicineId:
          type: integer
        medicineName:
          type: string
        status:
          type: string
          enum: [taken, skipped, missed]
        scheduledTime:
          type: string
          nullable: true
        takenAt:
          type: string
          format: date-time
      required: [id, medicineId, medicineName, status, takenAt]
...
    Recipe:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        ingredients:
          type: array
          items:
            type: string
        steps:
          type: array
          items:
            type: string
        calories:
          type: number
        proteinG:
          type: number
        carbsG:
          type: number
        fatG:
          type: number
        allergenFlags:
          type: array
          items:
            type: string
        cookTimeMinutes:
          type: integer
        servings:
          type: integer
      required: [title, description, ingredients, steps, calories, proteinG, carbsG, fatG, allergenFlags, cookTimeMinutes, servings]
...